// src/common/models/kaggle.ts
export enum KaggleUrls {
  LOGIN = "https://www.kaggle.com/account/login?phase=emailSignIn",
  DASHBOARD = "https://www.kaggle.com/",
  CSV_FILE = "https://www.kaggle.com/datasets/thedevastator/us-baby-names-by-year-of-birth?select=babyNamesUSYOB-full.csv"
};

export enum KaggleCssSelectors {
  LOGIN_EMAIL_INPUT = 'input[name="email"]',
  LOGIN_PASSWORD_INPUT = 'input[name="password"]',
  LOGIN_SUBMIT = 'button[type="submit"]',
  CSV_DOWNLOAD_BUTTON = 'span.google-symbols.notranslate.MuiIcon-root.MuiIcon-fontSizeMedium.sc-FFETS.bHohpu.notranslate.css-1jgtvd5',
  LOGOUT_BUTTON = 'a.sc-lgprfV.ZGfSk[aria-label="Sign Out"]',
  USER_MENU_BUTTON = 'div.sc-fDeYYK.gUohLu.sc-kUiVwB.fGsGmP'
};