// src/common/models/webForm.ts
export enum WebFormCssSelectors {
  YEAR_INPUT = 'input.whsOnd.zHQkBf[aria-labelledby="i1 i4"]',
  NAME_INPUT = 'input.whsOnd.zHQkBf[aria-labelledby="i6 i9"]',
  SEX_INPUT = 'input.whsOnd.zHQkBf[aria-labelledby="i11 i14"]',
  NUMBER_INPUT = 'input.whsOnd.zHQkBf[aria-labelledby="i16 i19"]',
  SUBMIT = 'div[role="button"]'
};